requirements.txt
numpy==1.20.3
wilds==1.2.2
imageio==2.9.0
gdown==3.13.0
torchvision==0.12.0
torch==1.11.0
tqdm==4.62.2
backpack==0.1
parameterized==0.8.1
Pillow
timm==0.4.12
POT=0.9.3